some work, still need to make animations
