+++ src/components/about/About.tsx
@@ -1,38 +1,12 @@
 import React, { useEffect, useRef, useState } from 'react';
 import 'animate.css';
 import './about.scss';
+import AnimatedBox from '../AnimatedBoxProps';
 
-const AnimatedBox: React.FC = () => {
-  const [isVisible, setIsVisible] = useState(false);
-  const animatedElement = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      if (animatedElement.current) {
-        const top = animatedElement.current.getBoundingClientRect().top;
-        const windowHeight = window.innerHeight;
-        setIsVisible(top < windowHeight);
-      }
-    };
-
-    window.addEventListener('scroll', handleScroll);
-    handleScroll();
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, []);
-
-  return (
-    <div ref={animatedElement} className={`box animate__animated ${isVisible ? 'animate__fadeIn animate__slower' : ''}`}><span className='about-title'>About Me</span></div>
-  );
-};
-
-const About: React.FC = () => {
-
+function About(){
   return (
     <div className='about-wrapper' id='about'>
-      <AnimatedBox />
+      <span className='about-title'>About Me</span>
       <div className='about-section'>
         <span className='about-p'>I'm a student at the Opole University of Technology in Poland. I'm currently studying Computer Science and I'm in my seccond year. I specialize in building websites, mainly focusing on frontend, but I also have no problem with backend.</span>
         <span className='about-p'>I love doing it because it brings me joy. I'm not afraid of challenging tasks, even though I know I don't have all the answers. I'm a fast learner and I'm always looking for new knowledge. I'm a team player, but I can also work alone. I prefer working among people, but I have no problem with remote work.</span>
